+++ src/pydvl/influence/torch/preconditioner.py
@@ -1,17 +1,20 @@
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from typing import Callable, Optional
+from typing import TYPE_CHECKING, Optional
 
 import torch
 
 from ...utils.exceptions import NotFittedException
-from .functional import LowRankProductRepresentation, randomized_nystroem_approximation
+from .functional import LowRankProductRepresentation, operator_nystroem_approximation
 
-__all__ = ["JacobiPreConditioner", "NystroemPreConditioner", "PreConditioner"]
+if TYPE_CHECKING:
+    from .operator import TensorOperator
+
+__all__ = ["JacobiPreconditioner", "NystroemPreconditioner", "Preconditioner"]
 
 
-class PreConditioner(ABC):
+class Preconditioner(ABC):
     r"""
     Abstract base class for implementing pre-conditioners for improving the convergence
     of CG for systems of the form
@@ -22,34 +25,57 @@ class PreConditioner(ABC):
     condition number than $A + \lambda \operatorname{I}$.
 
     """
+    _reg: Optional[float]
+
+    @property
+    def regularization(self):
+        return self._reg
+
+    @regularization.setter
+    def regularization(self, value: float):
+        if self.modify_regularization_requires_fit:
+            raise NotImplementedError(
+                f"Adapting regularization for instances of type "
+                f"{type(self)} without re-fitting is not "
+                f"supported. Call the fit method instead."
+            )
+        self._validate_regularization(value)
+        self._reg = value
+
+    def _validate_regularization(self, value: Optional[float]):
+        if value is not None and value < 0:
+            raise ValueError("regularization must be non-negative")
+
+    @property
+    @abstractmethod
+    def modify_regularization_requires_fit(self) -> bool:
+        pass
 
     @property
     @abstractmethod
     def is_fitted(self):
         pass
 
-    @abstractmethod
     def fit(
         self,
-        mat_mat_prod: Callable[[torch.Tensor], torch.Tensor],
-        size: int,
-        dtype: torch.dtype,
-        device: torch.device,
-        regularization: float = 0.0,
+        operator: "TensorOperator",
+        regularization: Optional[float] = None,
     ):
         r"""
         Implement this to fit the pre-conditioner to the matrix represented by the
         mat_mat_prod
         Args:
-            mat_mat_prod: a callable that computes the matrix-matrix product
-            size: size of the matrix represented by `mat_mat_prod`
-            dtype: data type of the matrix represented by `mat_mat_prod`
-            device: device of the matrix represented by `mat_mat_prod`
+            operator: The preconditioner is fitted to this operator
             regularization: regularization parameter $\lambda$ in the equation
                 $ ( A + \lambda \operatorname{I})x = \operatorname{rhs} $
         Returns:
             self
         """
+        self._validate_regularization(regularization)
+        return self._fit(operator, regularization)
+
+    @abstractmethod
+    def _fit(self, operator: "TensorOperator", regularization: Optional[float] = None):
         pass
 
     def solve(self, rhs: torch.Tensor):
@@ -73,12 +99,12 @@ class PreConditioner(ABC):
         pass
 
     @abstractmethod
-    def to(self, device: torch.device) -> PreConditioner:
+    def to(self, device: torch.device) -> Preconditioner:
         """Implement this to move the (potentially fitted) preconditioner to a
         specific device"""
 
 
-class JacobiPreConditioner(PreConditioner):
+class JacobiPreconditioner(Preconditioner):
     r"""
     Pre-conditioner for improving the convergence of CG for systems of the form
 
@@ -101,60 +127,70 @@ class JacobiPreConditioner(PreConditioner):
     """
 
     _diag: torch.Tensor
-    _reg: float
 
     def __init__(self, num_samples_estimator: int = 1):
         self.num_samples_estimator = num_samples_estimator
 
     @property
     def is_fitted(self):
-        return self._diag is not None and self._reg is not None
+        has_diag = hasattr(self, "_diag") and self._diag is not None
+        has_regularization = hasattr(self, "_reg")
+        return has_diag and has_regularization
 
-    def fit(
+    @property
+    def modify_regularization_requires_fit(self) -> bool:
+        return False
+
+    def _fit(
         self,
-        mat_mat_prod: Callable[[torch.Tensor], torch.Tensor],
-        size: int,
-        dtype: torch.dtype,
-        device: torch.device,
-        regularization: float = 0.0,
+        operator: "TensorOperator",
+        regularization: Optional[float] = None,
     ):
         r"""
         Fits by computing an estimate of the diagonal of the matrix represented by
-        `mat_mat_prod` via Hutchinson's estimator
+        a [TensorOperator][pydvl.influence.torch.base.TensorOperator]
+        via Hutchinson's estimator
 
         Args:
-            mat_mat_prod: a callable representing the matrix-matrix product
-            size: size of the square matrix
-            dtype: needed data type of inputs for the mat_mat_prod
-            device: needed device for inputs of mat_mat_prod
+            operator: The preconditioner is fitted to this operator
             regularization: regularization parameter
                 $\lambda$ in $(A+\lambda I)x=b$
         """
         random_samples = torch.randn(
-            size, self.num_samples_estimator, device=device, dtype=dtype
+            self.num_samples_estimator,
+            operator.input_size,
+            device=operator.device,
+            dtype=operator.dtype,
         )
+
         diagonal_estimate = torch.sum(
-            torch.mul(random_samples, mat_mat_prod(random_samples)), dim=1
+            torch.mul(random_samples, operator.apply(random_samples)), dim=0
         )
+
         diagonal_estimate /= self.num_samples_estimator
         self._diag = diagonal_estimate
         self._reg = regularization
 
     def _solve(self, rhs: torch.Tensor):
-        inv_diag = 1.0 / (self._diag + self._reg)
+        diag = self._diag
+
+        if self._reg is not None:
+            diag = diag + self._reg
+
+        inv_diag = 1.0 / diag
 
         if rhs.ndim == 1:
             return rhs * inv_diag
 
         return rhs * inv_diag.unsqueeze(-1)
 
-    def to(self, device: torch.device) -> JacobiPreConditioner:
+    def to(self, device: torch.device) -> JacobiPreconditioner:
         if self._diag is not None:
             self._diag = self._diag.to(device)
         return self
 
 
-class NystroemPreConditioner(PreConditioner):
+class NystroemPreconditioner(Preconditioner):
     r"""
     Pre-conditioner for improving the convergence of CG for systems of the form
 
@@ -177,7 +213,6 @@ class NystroemPreConditioner(PreConditioner):
     """
 
     _low_rank_approx: LowRankProductRepresentation
-    _regularization: float
 
     def __init__(self, rank: int):
         self._rank = rank
@@ -192,61 +227,61 @@ class NystroemPreConditioner(PreConditioner):
 
     @property
     def is_fitted(self):
-        return self._low_rank_approx is not None and self._regularization is not None
+        has_low_rank_approx = (
+            hasattr(self, "_low_rank_approx") and self._low_rank_approx is not None
+        )
+        has_regularization = hasattr(self, "_reg") and self._reg is not None
+        return has_low_rank_approx and has_regularization
 
-    def fit(
+    @property
+    def modify_regularization_requires_fit(self) -> bool:
+        return False
+
+    def _fit(
         self,
-        mat_mat_prod: Callable[[torch.Tensor], torch.Tensor],
-        size: int,
-        dtype: torch.dtype,
-        device: torch.device,
-        regularization: float = 0.0,
+        operator: "TensorOperator",
+        regularization: Optional[float] = None,
     ):
         r"""
         Fits by computing a low-rank approximation of the matrix represented by
         `mat_mat_prod` via Nystroem approximation
 
         Args:
-            mat_mat_prod: a callable representing the matrix-matrix product
-            size: size of the square matrix
-            dtype: needed data type of inputs for the mat_mat_prod
-            device: needed device for inputs of mat_mat_prod
+            operator: The preconditioner is fitted to this operator
             regularization: regularization parameter
                 $\lambda$  in $(A+\lambda I)x=b$
         """
 
-        self._low_rank_approx = randomized_nystroem_approximation(
-            mat_mat_prod, size, self._rank, dtype, mat_vec_device=device
-        )
-        self._regularization = regularization
+        self._low_rank_approx = operator_nystroem_approximation(operator, self._rank)
+        self._reg = regularization
 
     def _solve(self, rhs: torch.Tensor):
 
         rhs_is_one_dim = rhs.ndim == 1
+        b = torch.atleast_2d(rhs).t() if rhs_is_one_dim else rhs
 
-        rhs_view = torch.atleast_2d(rhs).t() if rhs_is_one_dim else rhs
+        U = self._low_rank_approx.projections
 
-        regularized_eigenvalues = (
-            self._low_rank_approx.eigen_vals + self._regularization
-        )
-        lambda_rank = self._low_rank_approx.eigen_vals[-1] + self._regularization
+        Sigma = self._low_rank_approx.eigen_vals
+        lambda_rank = self._low_rank_approx.eigen_vals[-1]
 
-        proj_rhs = self._low_rank_approx.projections.t() @ rhs_view
+        if self._reg is not None:
+            Sigma = Sigma + self._reg
+            lambda_rank = lambda_rank + self._reg
 
-        inverse_regularized_eigenvalues = lambda_rank / regularized_eigenvalues
+        U_t_b = U.t() @ b
 
-        result = self._low_rank_approx.projections @ (
-            proj_rhs * inverse_regularized_eigenvalues.unsqueeze(-1)
-        )
+        Sigma_inv = lambda_rank / Sigma
 
-        result += rhs_view - self._low_rank_approx.projections @ proj_rhs
+        result = U @ (U_t_b * Sigma_inv.unsqueeze(-1))
+        result += b - U @ U_t_b
 
         if rhs_is_one_dim:
             result = result.squeeze()
 
         return result
 
-    def to(self, device: torch.device) -> NystroemPreConditioner:
+    def to(self, device: torch.device) -> NystroemPreconditioner:
         if self._low_rank_approx is not None:
             self._low_rank_approx = self._low_rank_approx.to(device)
         return self