+++ tests/test_results.py
@@ -31,19 +31,17 @@ def dummy_values(values, names):
 )
 def test_sorting(values, names, ranks_asc, dummy_values):
     dummy_values.sort(key="value")
-    assert np.alltrue([it.value for it in dummy_values] == sorted(values))
-    assert np.alltrue(dummy_values.indices == ranks_asc)
-    assert np.alltrue(
-        [it.value for it in reversed(dummy_values)] == sorted(values, reverse=True)
-    )
+    assert [it.value for it in dummy_values] == sorted(values)
+    assert dummy_values.indices.tolist() == ranks_asc
+    assert [it.value for it in reversed(dummy_values)] == sorted(values, reverse=True)
 
     dummy_values.sort(reverse=True)
-    assert np.alltrue([it.value for it in dummy_values] == sorted(values, reverse=True))
-    assert np.alltrue(dummy_values.indices == list(reversed(ranks_asc)))
+    assert [it.value for it in dummy_values] == sorted(values, reverse=True)
+    assert dummy_values.indices.tolist() == list(reversed(ranks_asc))
 
     dummy_values.sort(key="index")
-    assert np.alltrue(dummy_values.indices == list(range(len(values))))
-    assert np.alltrue([it.value for it in dummy_values] == values)
+    assert dummy_values.indices.tolist() == list(range(len(values)))
+    assert [it.value for it in dummy_values] == values
 
 
 @pytest.mark.parametrize(
@@ -55,16 +53,16 @@ def test_dataframe_sorting(values, names, ranks_asc, dummy_values):
         import pandas
 
         df = dummy_values.to_dataframe(use_names=False)
-        assert np.alltrue(df.index.values == ranks_asc)
+        assert all(df.index.values == ranks_asc)
 
         df = dummy_values.to_dataframe(use_names=True)
-        assert np.alltrue(df.index.values == sorted_names)
-        assert np.alltrue(df["dummy_valuator"].values == sorted(values))
+        assert all(df.index.values == sorted_names)
+        assert all(df["dummy_valuator"].values == sorted(values))
 
         dummy_values.sort(reverse=True)
         df = dummy_values.to_dataframe(use_names=True)
-        assert np.alltrue(df.index.values == list(reversed(sorted_names)))
-        assert np.alltrue(df["dummy_valuator"].values == sorted(values, reverse=True))
+        assert all(df.index.values == list(reversed(sorted_names)))
+        assert all(df["dummy_valuator"].values == sorted(values, reverse=True))
     except ImportError:
         pass
 
@@ -87,15 +85,15 @@ def test_todataframe(ranks_asc, dummy_values):
     df = dummy_values.to_dataframe()
     assert "dummy_valuator" in df.columns
     assert "dummy_valuator_stderr" in df.columns
-    assert np.alltrue(df.index.values == ranks_asc)
+    assert all(df.index.values == ranks_asc)
 
     df = dummy_values.to_dataframe(column="val")
     assert "val" in df.columns
     assert "val_stderr" in df.columns
-    assert np.alltrue(df.index.values == ranks_asc)
+    assert all(df.index.values == ranks_asc)
 
     df = dummy_values.to_dataframe(use_names=True)
-    assert np.alltrue(df.index.values == [it.name for it in dummy_values])
+    assert all(df.index.values == [it.name for it in dummy_values])
 
 
 @pytest.mark.parametrize(
@@ -385,7 +383,7 @@ def test_adding_different_indices(
     [
         ([0, 1, 2], np.int32, ["a", "b", "c"], "<U1"),
         ([4, 1, 7], np.int64, [4, 1, 7], np.int64),
-        ([4, 1, 7], np.int32, [4, 1, 7], np.float_),
+        ([4, 1, 7], np.int32, [4, 1, 7], np.float64),
     ],
 )
 def test_types(indices, index_t, data_names, name_t):
@@ -394,7 +392,7 @@ def test_types(indices, index_t, data_names, name_t):
 
     v = ValuationResult(
         indices=np.array(indices, dtype=index_t),
-        values=np.ones(len(indices), dtype=np.float_),
+        values=np.ones(len(indices), dtype=np.float64),
         data_names=np.array(data_names, dtype=name_t),
     )
     assert v.indices.dtype == index_t