+++ CHANGELOG.md
@@ -30,8 +30,17 @@
 - Extend `NystroemSketchInfluence` with block-diagonal and Gauss-Newton
   approximation
   [PR #596](https://github.com/aai-institute/pyDVL/pull/596)
+- Extend `ArnoldiInfluence` with block-diagonal and Gauss-Newton
+  approximation
+  [PR #598](https://github.com/aai-institute/pyDVL/pull/598)
+- Extend `CgInfluence` with block-diagonal and Gauss-Newton
+  approximation
+  [PR #601](https://github.com/aai-institute/pyDVL/pull/601)
 
-### Fixed
+## Fixed
+- Replace `np.float_` with `np.float64` and `np.alltrue` with `np.all`,
+  as the old aliases are removed in NumPy 2.0
+  [PR #604](https://github.com/aai-institute/pyDVL/pull/604)
 
 - Fix a bug in pydvl.utils.numeric.random_subset where 1 - q was used instead of q
   as the probability of an element being sampled
@@ -61,7 +70,30 @@
     to `regularization` and change the type annotation to allow
     for block-wise regularization parameters
     [PR #596](https://github.com/aai-institute/pyDVL/pull/596)
-
+  - Renaming of parameters of `ArnoldiInfluence`,
+    `hessian_regularization` -> `regularization` (modify type annotation),
+    `rank_estimate` -> `rank`
+    [PR #598](https://github.com/aai-institute/pyDVL/pull/598)
+  - Remove functions remove obsolete functions 
+    `lanczos_low_rank_hessian_approximation`, `model_hessian_low_rank`
+    from `influence.torch.functional`
+    [PR #598](https://github.com/aai-institute/pyDVL/pull/598)
+  - Renaming of parameters of `CgInfluence`,
+    `hessian_regularization` -> `regularization` (modify type annotation),
+    `pre_conditioner` -> `preconditioner`,
+    `use_block_cg` -> `solve_simultaneously`
+    [PR #601](https://github.com/aai-institute/pyDVL/pull/601)
+  - Remove parameter `x0` from `CgInfluence`
+    [PR #601](https://github.com/aai-institute/pyDVL/pull/601)
+  - Rename module 
+    `influence.torch.pre_conditioner` -> `influence.torch.preconditioner`
+    [PR #601](https://github.com/aai-institute/pyDVL/pull/601)
+  - Refactor preconditioner:
+    - renaming `PreConditioner` -> `Preconditioner`
+    - fit to `TensorOperator`
+    [PR #601](https://github.com/aai-institute/pyDVL/pull/601)
+  
+  
 ## 0.9.2 - 🏗  Bug fixes, logging improvement
 
 ### Added